refactor: move batcher func to utils
