+++ src/libs/portfolio/portfolio.ts
@@ -10,8 +10,8 @@ import { PINNED_TOKENS } from '../../consts/pinnedTokens'
 import { Fetch } from '../../interfaces/fetch'
 import { Network } from '../../interfaces/network'
 import { RPCProvider } from '../../interfaces/provider'
+import batcher from '../../utils/batcher'
 import { Deployless, fromDescriptor } from '../deployless/deployless'
-import batcher from './batcher'
 import { isBlacklistedAsset, prepareBlacklistPatterns, STATIC_BLACKLIST } from './blacklist'
 import { portfolioDebugLog } from './debug'
 import { PORTFOLIO_LIB_ERROR_NAMES } from './errorNames'
